Add files via upload
Created a "death animation" which consists of death sprite becoming a trigger and falling through the stage. Messed around with and tested some other shit. The camera is kind of wonky so it doesn't really track well and also was playing with making Mario automatically die whenever the position of the Player object fell below a certain y-value.